ExtractorResult из Models поделены на отдельные файлы классов
