+++ Core/Models/Error.cs
@@ -1,0 +1,10 @@
+
+namespace ExtractorService.Models
+{
+    public class Error
+    {
+        public string Reason { get; set; }
+        public int idLogs { get; set; }
+        public Enum PlaceType { get; set; } 
+    }
+}
