+++ Core/Models/ErrorResultInfo.cs
@@ -1,0 +1,8 @@
+
+namespace ExtractorService.Models{
+    public class ErrorResultInfo
+    {
+        public List<Error> Errors { get; set; }
+        public int ErrorCount { get; set; }
+    }
+}
