+++ Core/Models/ExtractorSettings.cs
@@ -1,0 +1,10 @@
+
+namespace ExtractorService.Models
+{
+    public class ExtractorSetting
+    {
+        public string URL { get; set; }
+        public ExtractorType Type { get; set; }
+        public int BatchCount { get; set; }
+    }
+}
