+++ Core/Models/ExtractorType.cs
@@ -1,0 +1,9 @@
+
+namespace ExtractorService.Models
+{
+    public class ExtractorType
+    {
+        public Enum SourceType { get; set; }
+        public ExtractInputData InputData { get; set; }
+    }
+}
