+++ Core/Models/PlaceType.cs
@@ -1,0 +1,11 @@
+
+
+namespace ExtractorService.Models
+{
+    public enum PlaceType
+    {
+        Extracting = 0,
+        Handling = 1,
+        Saving = 2
+    }
+}
