+++ Core/Models/SourceType.cs
@@ -1,0 +1,10 @@
+
+namespace ExtractorService.Models
+{
+    public enum SourceType 
+    {
+        API = 0,
+        HTML = 1,
+        SCV = 2
+    }
+}
